# backend/src/main/kotlin/io/github/projektalmanac/imani/generated/api/CitaApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package io.github.projektalmanac.imani.generated.api

import io.github.projektalmanac.imani.generated.dto.CitasDto
import io.github.projektalmanac.imani.generated.dto.NuevaCitaDto
import io.github.projektalmanac.imani.generated.dto.ProblemDetailsDto
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Email
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size
import javax.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@Validated
interface CitaApi {

    @Operation(
        tags = ["cita",],
        summary = "Obtiene las Citas del paciente",
        operationId = "getDates",
        description = """Obtienen la cita del paciente con la informacion del doctor y centro medico""",
        responses = [
            ApiResponse(responseCode = "200", description = "OK", content = [Content(array = ArraySchema(schema = Schema(implementation = CitasDto::class)))]),
            ApiResponse(responseCode = "404", description = "Paciente no encontrado", content = [Content(schema = Schema(implementation = ProblemDetailsDto::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/pacientes/{pacienteId}/citas"],
            produces = ["application/json"]
    )
    fun getDates(@Parameter(description = "", required = true) @PathVariable("pacienteId") pacienteId: kotlin.Int): ResponseEntity<List<CitasDto>>

    @Operation(
        tags = ["cita",],
        summary = "Crea cita medica",
        operationId = "postDate",
        description = """Crea una cita medica a un paciente se envia id Doctor y dia de la cita""",
        responses = [
            ApiResponse(responseCode = "201", description = "cita creada"),
            ApiResponse(responseCode = "404", description = "Paciente no encontrado", content = [Content(schema = Schema(implementation = ProblemDetailsDto::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/pacientes/{pacienteId}/citas"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun postDate(@Parameter(description = "", required = true) @PathVariable("pacienteId") pacienteId: kotlin.Int,@Parameter(description = "") @Valid @RequestBody(required = false) nuevaCitaDto: NuevaCitaDto?): ResponseEntity<Unit>
}