# backend/src/main/kotlin/io/github/projektalmanac/imani/generated/api/PrescripcionesApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.7.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package io.github.projektalmanac.imani.generated.api

import io.github.projektalmanac.imani.generated.dto.NuevaPrescripcionDto
import io.github.projektalmanac.imani.generated.dto.PrescripcionDto
import io.github.projektalmanac.imani.generated.dto.ProblemDetailsDto
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import javax.validation.constraints.DecimalMax
import javax.validation.constraints.DecimalMin
import javax.validation.constraints.Email
import javax.validation.constraints.Max
import javax.validation.constraints.Min
import javax.validation.constraints.NotNull
import javax.validation.constraints.Pattern
import javax.validation.constraints.Size
import javax.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@Validated
interface PrescripcionesApi {

    @Operation(
        tags = ["prescripciones",],
        summary = "Obtiene las prescripciones del paciente",
        operationId = "getPrescription",
        description = """Medicamentos, horas, inicio y duracion""",
        responses = [
            ApiResponse(responseCode = "200", description = "OK", content = [Content(array = ArraySchema(schema = Schema(implementation = PrescripcionDto::class)))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/pacientes/{pacienteId}/prescripciones"],
            produces = ["application/json"]
    )
    fun getPrescription(@Parameter(description = "", required = true) @PathVariable("pacienteId") pacienteId: kotlin.Int): ResponseEntity<List<PrescripcionDto>>

    @Operation(
        tags = ["prescripciones",],
        summary = "Agregar prescripciones",
        operationId = "postPrescription",
        description = """agrega una lista de prescripciones (Medicamentos, horas, inicio y duracion)""",
        responses = [
            ApiResponse(responseCode = "201", description = "Created"),
            ApiResponse(responseCode = "404", description = "Paciente no encontrado", content = [Content(schema = Schema(implementation = ProblemDetailsDto::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/pacientes/{pacienteId}/prescripciones"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun postPrescription(@Parameter(description = "", required = true) @PathVariable("pacienteId") pacienteId: kotlin.Int,@Parameter(description = "") @Valid @RequestBody(required = false) nuevaPrescripcionDto: kotlin.collections.List<NuevaPrescripcionDto>?): ResponseEntity<Unit>

    @Operation(
        tags = ["prescripciones",],
        summary = "Modifica Prescripcion",
        operationId = "putPrescripcion",
        description = """Modifica la prescripcion de un medicamento""",
        responses = [
            ApiResponse(responseCode = "204", description = "No Content"),
            ApiResponse(responseCode = "400", description = "Bad Request", content = [Content(schema = Schema(implementation = ProblemDetailsDto::class))]),
            ApiResponse(responseCode = "404", description = "Not Found", content = [Content(schema = Schema(implementation = ProblemDetailsDto::class))])
        ]
    )
    @RequestMapping(
            method = [RequestMethod.PUT],
            value = ["/pacientes/{pacienteId}/prescripciones/{prescripcionId}"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun putPrescripcion(@Parameter(description = "", required = true) @PathVariable("pacienteId") pacienteId: kotlin.Int,@Parameter(description = "", required = true) @PathVariable("prescripcionId") prescripcionId: kotlin.Int,@Parameter(description = "") @Valid @RequestBody(required = false) nuevaPrescripcionDto: NuevaPrescripcionDto?): ResponseEntity<Unit>
}