# front-medico/src/generated/openapi/model/nuevoFarmaceutico.ts
/**
 * Imani API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface NuevoFarmaceutico { 
    nombre: string;
    apellidos: string;
    nombreUsuario: string;
    password: string;
}